refactor(playlists): split programmaticPlaylists service into per-mix modules
Relocate the 4,003-line service into services/programmaticPlaylists/
as an inheritance chain over a shared base: activity, audio-analysis,
contextual, curated, library, and weekly/mood mix generators plus
shared types and constants, composed into the same public
ProgrammaticPlaylistService and singleton. The façade re-exports the
identical runtime and type surface (proven old-vs-new). Generator
bodies moved text-identical; the only access-modifier widening is
private to protected where inheritance requires it. No module-level
mutable state exists on either side; ratchet baselines untouched.

Adds a module-boundary behavioral suite asserting the export list,
singleton identity, and representative behavior from every module
through the façade.

Refs #485
