+++ backend/src/services/programmaticPlaylists/index.ts
@@ -1,0 +1,12 @@
+export {
+    ProgrammaticPlaylistService,
+    programmaticPlaylistService,
+} from "./service";
+export type { ProgrammaticMix } from "./shared";
+export { seededShuffle } from "../artistSlotAllocation";
+export {
+    applyArtistCap,
+    type ArtistCapTrack,
+    type ArtistCapFallbackOptions,
+    type ApplyArtistCapOptions,
+} from "../programmaticPlaylistArtistCap";
